+++ src/mir/caching/legendre/FileLoader.cc
@@ -28,7 +28,7 @@ namespace legendre {
 
 FileLoader::FileLoader(const param::MIRParametrisation& parametrisation, const eckit::PathName& path) :
     LegendreLoader(parametrisation, path),
-    buffer_(path.size()) {
+    buffer_(size_t(path.size())) {
 
     log() << "Loading Legendre coefficients from " << path << std::endl;
