+++ src/mir/caching/matrix/FileLoader.cc
@@ -27,7 +27,7 @@ namespace matrix {
 
 FileLoader::FileLoader(const std::string& name, const eckit::PathName& path) :
     MatrixLoader(name, path),
-    buffer_(path.size()) {
+    buffer_(size_t(path.size())) {
 
     log() << "Loading matrix from " << path << std::endl;
